Functions for absorption spectrum calculcation

The most important function of the module is the α function. Before using i, you have to initialise a database and use the fetch! function to retrieve line-by-line data.

The absorption coefficient

HITRAN.α — Function

α(tables::AbstractVector{String} [, profile=:hartmann_tran; kwargs...])

Computes the absorption coefficient using line-by-line data stored in the database tables specified in tables. The lineshape can be optionally specified using the profile argument and one of the Symbol keys :hartmann_tran, :voigt, :sdvoigt, :lorentz, :gauss. If no keyword arguments are specified, they will be automatically chosen from the tables provided.

Keyword arguments

components: the components of the gas mixture for the calculation. Can be either a vector of tuples with (molecule_id, local_iso_id) or a Dict with the (molecule_id, local_iso_id) tuple as key and the abundance as value. If the vector of tuples is supplied the natural abundance will be used, so this makes no sense for gas mixtures other than isotopologues of the same molecule.

intensity_threshold: the minimum line strength in $cm^{-1}/(\text{molecule} \cdot cm^{-2})$

pressure: the environmental pressure in atmospheres (default: 1.0 atm)

temperature: the environmental temperature in Kelvin (default: 296.0 K)

ν_range: a tuple of the form (νmin, νmax) where νmin/νmax is the minimum/maximum wavenumber for the absorption_spectrum respectively in $cm^{-1}$

ν_step: the wavenumber step in $cm^{-1}$ (default: 0.01 $cm^{-1}$)

ν_wing: absolute calculation width of a line in $cm^{-1}$ (default: 0 $cm^{-1}$)

ν_wing_hw: relative calculation width of a line in multiples of a half-width (default: 50)

diluent: a Dict of the diluting substances, specified as Symbol, e.g. :air or :H2O for the key and the relative concentration as key (default: Dict(:self => 1.0)). See the example for details.

Environments

There are two default environments defined at the moment, which ease the calculation of atmospheric spectra.

Dry air

The first environment is dry air which can be accessed using default_environments[:dry_air]. The following composition is used [Picard2008]:

Gas	Volume (ppmv)

$N_2$	780,848

$O_2$	209,390

$Ar$	9,332

$CO_2$	400

$Ne$	18.2

$CH_4$	1.5

$Kr$	1.1

$H_2$	0.5

$N_2O$	0.3

$CO$	0.2

$Xe$	0.1

Moist air

The second environment model is moist air which takes the relative humidity into the account. You can use the function moist_air to get a composition dictionary with the correct water concentration.

Convenience functions

The following three functions are provided as a convenience to convert the absorption coefficient to either an absorption spectrum, transmittance spectrum or an optical depth.

HITRAN.absorption_spectrum — Method

absorption_spectrum(α::AbstractVector, len)

Computes the absorption spectrum for the given length len in centimeters. The vector α should be calculated using the corresponding α function.

HITRAN.transmittance_spectrum — Method

transmittance_spectrum(α::AbstractVector, len)

Computes the transmittance spectrum for the given length len in centimeters. The vector α should be calculated using the corresponding α function.

HITRAN.optical_depth — Method

optical_depth(α::AbstractVector, len)

Computes the optical depth for the given length len in centimeters. The vector α should be calculated using the corresponding α function.

Picard2008A., Picard, R.S., Davis, M., Gläser and K., Fujii (2008), Revised formula for the density of moist air (CIPM-2007), Metrologia 45, 149–155 (2008). Link to article